backend/pyproject.toml
[project]
name = "orgst"
version = "0.1.0"
description = "django-ninja project linters template"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.2",
    "django-environ>=0.12.1",
    "django-ninja>=1.5.3",
    "psycopg>=3.3.2",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "pylint>=4.0.4",
    "pylint-django>=2.7.0",
    "pytest>=9.0.2",
    "pytest-django>=4.11.1",
    "ruff>=0.15.1",
]
